requirements.txt
setuptools==46.1.3
pandas==1.0.3
pillow==7.1.2